rag_pdf-master/requirements.txt
langchain>=0.1.13
langchain-community>=0.0.32
langchain-core>=0.1.42
langchain-text-splitters>=0.0.1
requests>=2.31.0
huggingface-hub>=0.21.4
sentence-transformers>=2.7.0
chromadb>=0.4.24
streamlit>=1.33.0
python-dotenv>=1.0.0
pypdf>=3.17.0
docx2txt>=0.8
unstructured[local-inference]>=0.12.2
openpyxl>=3.1.2
python-magic>=0.4.27
docx2txt==0.8